move jwt methode to participantController
